requirements.txt
pytorch-lightning==1.3.8
torchvision==0.6.0
pytorch==1.5.0
opencv-contrib-python-headless>=4.1.2.30
matplotlib>=3.0.2
pandas>=0.23.3
scikit_learn>=0.22.1
promise>=2.2.1
slackclient==1.3.0
PyYAML==5.1
redis==2.10.6
numpy>=1.18.1
jsonschema==2.6.0
dill==0.2.8.2
efficientnet-pytorch>=0.6.1
albumentations>=0.4.3
tensorboard==2.2.0
thop>=0.0.31